requirements.txt
discord.py
python-dotenv
aiosqlite
pillow
requests
